+++ packages/cli/moondebug.mjs
@@ -1,0 +1,3 @@
+#!/usr/bin/env -S node --inspect
+
+import './dist/cmds/entrypoint.js'