+++ packages/cli/src/cmds/runTests.ts
@@ -32,6 +32,11 @@ export async function testCmd(envName: string, additionalArgs?: object): Promise
   ) {
     clearNodeLogs();
   }
+
+  if (env.foundation.type === "zombie") {
+    process.env.MOON_EXIT = "true";
+  }
+
   const vitest = await executeTests(env, additionalArgs);
   const failed = vitest!.state.getFiles().filter((file) => file.result!.state === "fail");
 
@@ -68,9 +73,10 @@ export async function executeTests(env: Environment, additionalArgs?: object) {
         const { rtVersion, rtName } = Object.values(chainData[0])[0];
         process.env.MOON_RTVERSION = rtVersion;
         process.env.MOON_RTNAME = rtName;
+      } catch (e) {
+        console.error(e);
+      } finally {
         await MoonwallContext.destroy();
-      } catch {
-        // No chain to test against
       }
     }
 
@@ -124,6 +130,7 @@ function addThreadConfig(
 ): UserConfig {
   const configWithThreads: UserConfig = {
     ...config,
+    fileParallelism: false,
     pool: "threads",
     poolOptions: {
       threads: {
@@ -131,12 +138,13 @@ function addThreadConfig(
         minThreads: 1,
         maxThreads: 1,
         singleThread: false,
-        useAtomics: false,
+        useAtomics: true,
       },
     },
   };
 
   if (threads == true && process.env.MOON_RECYCLE !== "true") {
+    configWithThreads.fileParallelism = true;
     configWithThreads.poolOptions.threads = {
       isolate: true,
       minThreads: 1,
@@ -146,12 +154,13 @@ function addThreadConfig(
     };
   }
 
-  if (typeof threads === "number") {
+  if (typeof threads === "number" && process.env.MOON_RECYCLE !== "true") {
+    configWithThreads.fileParallelism = true;
     configWithThreads.poolOptions.threads.maxThreads = threads;
     configWithThreads.poolOptions.threads.singleThread = false;
   }
 
-  if (typeof threads === "object") {
+  if (typeof threads === "object" && process.env.MOON_RECYCLE !== "true") {
     const key = Object.keys(threads)[0];
     if (["threads", "forks", "vmThreads", "typescript"].includes(key)) {
       configWithThreads.pool = key as "threads" | "forks" | "vmThreads" | "typescript";