+++ packages/cli/src/internal/nodePool.ts
@@ -1,0 +1,18 @@
+// Define as a forked resource
+
+/// Use acquire use release
+
+// HTTP Server example: https://github.com/mikearnaldi/sf-meetup/blob/main/src/lib/http.ts
+
+// Effect.forever,
+// Effect.interruptible,
+// Effect.forkScoped
+
+// Use message queues: https://effect.website/docs/concurrency/queue
+
+// IPC Server!
+
+// Needs endpoints:
+// request
+// release
+// query