+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @moonwall/types
 
+## 4.5.1
+
+## 4.5.0
+
+## 4.4.5
+
+### Patch Changes
+
+- 14d65d8: Another Concurrency Fix
+
+## 4.4.4
+
+## 4.4.3
+
+### Patch Changes
+
+- efe7ba1: Concurrency Fix
+
 ## 4.4.2
 
 ### Patch Changes